pyproject.toml
[tool.poetry]
name = "data-gpt"
version = "0.1.0"
description = ""
authors = ["Eirik Folkestad <eirik.ek.folkestad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
guidance = "^0.0.64"
transformers = "^4.30.2"
python-dotenv = "^1.0.0"
duckdb = "^0.8.1"
pandas = "^2.0.2"
google-cloud-bigquery = "^3.11.1"
db-dtypes = "^1.1.1"
chromadb = "^0.3.26"
rich = "^13.4.2"
click = "^8.1.3"
slack-bolt = "^1.18.0"
columnar = "^1.4.1"
pyshorteners = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.272"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
